backend/requirements.txt
# Core dependencies for Vercel deployment
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
sqlalchemy>=2.0.36
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
psycopg[binary]>=3.2.0
python-multipart>=0.0.12
mangum>=0.19.0
httpx>=0.28.0
